feat: read mailbox every 15 minutes
